+++ src/Sprite.cpp
@@ -5,14 +5,14 @@
 
 #include "Graphics.h"
 
-Sprite::Sprite(Graphics &graphics, const std::string &spritePath, int positionX, int positionY) :
+Sprite::Sprite(Graphics &graphics, const std::string &spritePath, double positionX, double positionY) :
         spriteSheet(graphics.createTexture(spritePath)),
         sourceRect{0, 0, 0, 0},
         position{positionX, positionY}
 {}
 
 Sprite::Sprite(Graphics &graphics, const std::string &spritePath, int sourceX, int sourceY,
-               int sourceW, int sourceH, int positionX, int positionY) :
+               int sourceW, int sourceH, double positionX, double positionY) :
         spriteSheet(graphics.createTexture(spritePath)),
         sourceRect{sourceX, sourceY, sourceW, sourceH},
         position{positionX, positionY}
@@ -23,7 +23,7 @@ Sprite::~Sprite() {
 }
 
 void Sprite::draw(Graphics &graphics) {
-    SDL_Rect destRect{std::lround(position.x), std::lround(position.y), (int)(this->sourceRect.w * globals::spriteScale),
-                      (int)(this->sourceRect.h * globals::spriteScale)};
+    SDL_Rect destRect{static_cast<int>(std::lround(position.x)), static_cast<int>(std::lround(position.y)), static_cast<int>(this->sourceRect.w * globals::spriteScale),
+                      static_cast<int>(this->sourceRect.h * globals::spriteScale)};
     graphics.blitSurface(*this->spriteSheet, this->sourceRect, destRect);
 }
